requirements.txt
Flask==3.1.2
Flask-PyMongo==3.0.1
flask-mongoengine @ git+https://github.com/idoshr/flask-mongoengine.git@e244408acf440c4208f7ddcd6e5d819cb472e4da
bson==0.5.10